# refactor: use klauspost/compress for compression