Use more performant optimized packages for compression: [github.com/klauspost/compress](https://github.com/klauspost/compress)

Benchmark results:

old:

```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sumologicexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkCompression
BenchmarkCompression/deflate
BenchmarkCompression/deflate-16                   203900             17899 ns/op           86267 B/op         20 allocs/op
BenchmarkCompression/gzip
BenchmarkCompression/gzip-16                      175705             20847 ns/op           87679 B/op         22 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sumologicexporter    31.009s
Execution time: 0h:00m:32s sec                                                                                                                                                                                                                                
```

new:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sumologicexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkCompression
BenchmarkCompression/deflate
BenchmarkCompression/deflate-16                   140374             26205 ns/op           86265 B/op         20 allocs/op
BenchmarkCompression/gzip
BenchmarkCompression/gzip-16                       48931             68006 ns/op           87689 B/op         22 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sumologicexporter    20.145s
```

benchstat:

```                                            
name                    old time/op    new time/op    delta
Compression/deflate-16    26.2µs ± 0%    17.9µs ± 0%  -31.70%
Compression/gzip-16       68.0µs ± 0%    20.8µs ± 0%  -69.35%

name                    old alloc/op   new alloc/op   delta
Compression/deflate-16    86.3kB ± 0%    86.3kB ± 0%   +0.00%
Compression/gzip-16       87.7kB ± 0%    87.7kB ± 0%   -0.01%

name                    old allocs/op  new allocs/op  delta
Compression/deflate-16      20.0 ± 0%      20.0 ± 0%    0.00%
Compression/gzip-16         22.0 ± 0%      22.0 ± 0%    0.00%
```