@pmalek-sumo 2021-07-05
> Are we fine with CGO?

This package doesn't seem to leverage cgo:

```
➜  ~/code/go git clone https://github.com/klauspost/compress.git
Cloning into 'compress'...
remote: Enumerating objects: 5495, done.
remote: Counting objects: 100% (460/460), done.
remote: Compressing objects: 100% (221/221), done.
remote: Total 5495 (delta 257), reused 408 (delta 229), pack-reused 5035
Receiving objects: 100% (5495/5495), 37.02 MiB | 19.77 MiB/s, done.
Resolving deltas: 100% (2899/2899), done.
➜  ~/code/go cd compress
➜  ~/code/go/compress git:(master) find . -iname "*cgo*"
➜  ~/code/go/compress git:(master)
```