@pmalek-sumo 2021-07-05
> ```
> The raw results are in my updated spreadsheet. Due to cgo changes and upstream updates i could not get the cgo version of gzip to compile. Instead I included the zstd cgo implementation. If I get cgo gzip to work again, I might replace the results in the sheet.
> ```

That confirms what I wrote: there are no cgo files as of now in the repo -> regardless of `CGO_ENABLED` env, all code uses native go code.