@sumo-drosiek 2021-07-06
I don't know. This has been rejected from opentelemetry repo due to CGO :shrug: